requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.14.0
et-xmlfile==1.1.0
openpyxl==3.0.10
psycopg2==2.9.5
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.6
